Fix OkHttpConnection nullpointerexception when parsing content type
